+++ packages/agent/src/agent/http/index.ts
@@ -4,10 +4,11 @@ import { AgentError } from '../../errors';
 import { AnonymousIdentity, Identity } from '../../auth';
 import * as cbor from '../../cbor';
 import { RequestId, hashOfMap, requestIdOf } from '../../request_id';
-import { bufFromBufLike, concat, fromHex } from '../../utils/buffer';
+import { bufFromBufLike, concat, fromHex, toHex } from '../../utils/buffer';
 import {
   Agent,
   ApiQueryResponse,
+  HttpDetailsResponse,
   QueryFields,
   QueryResponse,
   ReadStateOptions,
@@ -28,7 +29,7 @@ import {
   ReadRequestType,
   SubmitRequestType,
 } from './types';
-import { AgentHTTPResponseError } from './errors';
+import { AgentCallError, AgentHTTPResponseError } from './errors';
 import { SubnetStatus, request } from '../../canisterStatus';
 import {
   CertificateVerificationError,
@@ -512,6 +513,7 @@ export class HttpAgent implements Agent {
 
     const body = cbor.encode(transformedRequest.body);
     const backoff = this.#backoffStrategy();
+    const requestId = requestIdOf(submit);
     try {
       // Attempt v3 sync call
       const requestSync = () => {
@@ -543,7 +545,6 @@ export class HttpAgent implements Agent {
         backoff,
         tries: 0,
       });
-      const requestId = requestIdOf(submit);
 
       const response = await request;
       const responseBuffer = await response.arrayBuffer();
@@ -586,9 +587,19 @@ export class HttpAgent implements Agent {
           identity,
         );
       }
-
-      this.log.error('Error while making call:', error as AgentError);
-      throw error;
+      const callError = new AgentCallError(
+        'Encountered an error while making call:',
+        error as HttpDetailsResponse,
+        toHex(requestId),
+        toHex(transformedRequest.body.sender_pubkey),
+        toHex(transformedRequest.body.sender_sig),
+        String(transformedRequest.body.content.ingress_expiry['_value']),
+      );
+      this.log.error(
+        `Error while making call: ${(error as Error).message ?? String(error)}`,
+        callError,
+      );
+      throw callError;
     }
   }